Le zoomOutElastique positionne correctement la fenetre
Ne corrige pas le rapport d'aspect pour l'instant, mais la position
est correcte.
Changement de l'interface native et de ProjectionOrtho.h afin de
diminuer les conversions ; maintenant des double sont les types
acceptés lors de l'appel à zoomOut.

+++ Commun/Utilitaire/Vue/ProjectionOrtho.h
@@ -54,7 +54,7 @@ namespace vue {
 
 
 		/// Zoom out élastique, sur un rectangle.
-		void zoomerOut(const glm::ivec2& coin1, const glm::ivec2& coin2);
+		void zoomerOut(const glm::dvec2& coin1, const glm::dvec2& coin2);
 		/// Zoom in élatique, sur un rectangle.
 		void zoomerIn(const glm::ivec2& coin1, const glm::ivec2& coin2);
